webapp: apps can assume an IAM role
Ability to attach an IAM role to the app so that we can control the level
of permissions it has (e.g. access to S3 buckets, etc...)

The IAM role name is passed via the `AWS.IAMRole` chart value and it's
optional.

Under the hood the `iam.amazonaws.com/role` annotation is used by the
[kube2iam](https://github.com/jtblin/kube2iam) daemonset running in our
kubernetes cluster.