FluentD Temp Fork
Trello: https://trello.com/c/W5NVPEQ0

After upgrading K8s to 1.8.9 we fell victim to
https://github.com/kubernetes/kubernetes/pull/58720

Ideally the fluentd container should source values from env vars instead
of writing them to a config file. See:
[entryPoint](https://github.com/fluent/fluentd-kubernetes-daemonset/blob/master/docker-image/v0.12/debian-elasticsearch/entrypoint.sh)

The `set -e` does not help either.

The resource change adresses the logs mentioned in the [trello](https://trello.com/c/W5NVPEQ0) ticket.

Ruby (Particularly in Alpha) is throwing OOM events as it exceeds it's
limits of 256MB and sometimes spikes well above.  When this occurs the
Kubelet daemon kills the pod and recreates it (As it is supposed to).
However, this event happens so often that the `kubelet` and `docker`
daemon are not able to keep up with clearing up the network sanbox and this
causes CNI errors to occur.

**Deployed** in alpha and dev