rstudio 1.4.2: Sessions last 8 hours
The RStudio auth-proxy session cookies expired after 1h but as users keep
they RStudio tab open for longer and because of how RStudio works (it's
essentially a single page app not making new HTTP requests).

Users get a `unable to establish a connection with the R session` when
trying to interact with RStudio console once session expired.

Session now lasts 8 hours.

Ticket: https://trello.com/c/CP1tWlzK/726-2-rstudio-unable-to-establish-a-connection-with-the-r-session-every-hour