Improve logic of script which updates helm chart repository
Old logic/problem
=================

The existing logic to determine if a package for an helm chart
needed update was comparing the time of the last commit for the
helm chart with the time of the last commit of the index file.

This was mostly working but not always. It wasn't updating an helm
chart package when another PR was merged before.
In this case the index was rebuilt and therefore commits older than
this time were not causing the re-build of the index/packages.

New logic
=========

I've tweaked the `update-packages` script to compare the time when
the chart code was last updated with the time one of the charts' tarballs
was last updated.

This means that each chart package is treated independently and the risk
of non-update is limited.

Caveat (spelled right this time)
================================

Potentially non-updates are still possible if there multiple PRs for
the same helm chart. This is obviously less likely.