Made set_rstudio_image_tag.sh script more explicit
Before someone tries to run it to update the auth-proxy image and
accidentally updates the rstudio docker image.

Previously the assumption was that app selector and docker image name
were the same (`rstudio`). This is true in the case of the main
container but not in the case of the auth-proxy container (
whose image is `rstudio-auth-proxy`).