airflow-k8s: Volumes are working now
This is mainly about having the NFS volume configured and wired
correctly.

Other improvements:
- Fixed `kubernetes.worker_container_*` settings to use right image
- Fixed `kubernetes.namespace` to use helm release namespace

Notes on volumes
================

Because of a [bug in Airflow+Kubernetes' `WorkerConfiguration`](https://github.com/apache/incubator-airflow/commit/e30a1f451aa5ec5aca4c886067ba8946a3d33395#r29462135)
the `kubernetes.(dags|logs)_volume_subpath` configurations don't work
as they should (it's setting the `subPath` property on the volume but
this is a property of `volumeMounts`).
This means we can't use the same claim for both DAGs/logs at the moment.

The workaround for the above is that we create 3 PV:
- `storage` for the whole NFS airflow volume
- `dags` and `logs` pointing to the `/dags`/`/logs` path of the same NFS
  volume above

We then create 3 PVC:
- `storage`, `dags` and `logs` pointing to the corresponding PVs described
   above

Then in our deployment/containers we only refer to the `storage` one (
leveraging the `subPath` property of `volumeMounts`).

Although the `*-dags` and `*-logs` PVC are not used directly by us, they're
used by Airflow (see `(dags|logs)_volume_claim` setting in Airflow
configuration/configmap.