Airflow: allow getting new dags via a git repo
This commit stops copying in the example dags from the airflow distribution and
instead pulls them in from a repo on pod start. This isn't production ready but
it allows us to test Airflow in a similar way to how we envision users would.

We can merge new DAGs into our example repo, restart the pod and the new DAG
will be available. I use curl to pull the tarball from github instead of using a
git-sync sidecar, I tried this approach but ran into permission issues on the
nfs share and the expectation of the git-sync container to clone into a
directory tree it fully controls. This can be made to work but adds too much
complexity. It's something we'll have to tackle if we go to production with
airflow.