Better Probes 3
Modifying `livenessProbe` and `readinessProbe` to type `exec` for Jupyter container

Since introducing probes for Jupyter we have intermittently witnessed
instability when starting the `pod`.

At present we're still unsure as the root cause of the probes failing
for a seemingly healthy container.  The response from Jupyter to the probes on
some occassions appears to be invalid. Initially the probes were introduced
to address failures with the authentication proxy container and they
appear to be working.