Remove Jupyter Probes
Since last chart update stability has continued to decrese due what
appears to be a command in the Dockerfile taking a long time.  The base
image was updated 2 months ago and may have been fixed since.

This PR is on standby in case we have not confirmed a fix by COB.