JupyterLab v0.1.14: Return of the probes
The probes are back now the underlying problem causing slow jupyter startup has
been identified and fixed. The fix was to run the jupyter container as uid=1001
so it matches rstudio. We were doing this in the docker file before using the
NB_UID environment variable but the presence of that caused the underlying
jupyter base image to `chown` the users' home directory. The bigger the home
directory, the slower this made startup.