images-prepuller: Pre-pull/caches images on kubernetes nodes
Why?
=====================================================================

Some of our Docker images are quite big and it can take some time for
them to be downloaded.

If that image was not already downloaded on that node, the first user
which spawns a tool using that docker image will have to wait more than
the usual.

This can be confusing for the users as most of the time these
tools starts relatively quickly. When this takes longer the user
perceive this delay as a symptom of "something didn't work".

This is made worse by the fact that we now have configured our
kubernetes cluster to scale up and down. This means that new
nodes could be created more often. These fresh nodes
will not have the most used Docker images yet.

How does it work?
=====================================================================

This helm chart create a DaemonSet (ds) called `prepull`.
It will start a pod with an init container on each of the nodes.
This init container will go through the list of docker images to
cache and will run a `docker pull $IMAGE`.

**NOTE**: As these init containers are, well, containers we need
to mount the node's path where the Docker images are stored
(`/var/run`) in order for the docker images' files to not be
lost when this container terminate.