airflow-k8s 0.0.9: Fixed failing init job when re-installing
Chart installation when chart was previously installed and purged.
This was caused by `init-airflow-ns` job that annotate the
`airflow` namespace.

Techical details: The problem was that the `Job` created on new
install (`else` statement) was not idempotent because it was
missing the `--overwrite` flag (which the `if` block had).

I've simplified the init job template to factor all the
common parts and have a smaller `if` just for the
`helm.sh/hook` annotation.