Pod Priority Classess
[Trello](https://trello.com/c/Ay2ryNvC)

Adding a priority class to be referenced by system pods

We want some pods (particularly those that handle log and metrics aggregation) to be considered a higher priority by the scheduler.
We had an instance where a node was experiencing resource contention and
a `daemonset` controlled pod was left in a `pending` state after the pod
restarted.  `Daemonset` pods have the `nodeName` field set so if the node specified has no room to
accomodate the workload then the Scheduler has no choice but to wait for
the specified node to become available.

A pod referencing the `system` priority class will always be scheduled.
Even if it means evicting other lower priority pods. Pods that have no
priority set have a priority of `0`.