rstudio: Attempt to use generic auth-proxy
RStudio is the only application not using our [generic auth-proxy].

RStudio was using a special proxy which set an authentication header.
This commit attempts to disable the RStudio authentication.
By doing so we can use the plain auth proxy instead.

NOTE: The configuration for the proxy was copied and tweaked from the
jupyter-lab helm chart.

[generic auth-proxy]: https://github.com/ministryofjustice/analytics-platform-auth-proxy

Ticket: https://trello.com/c/5twZHrb4/215-try-to-replace-custom-rstudio-auth-proxy-with-our-usual-generic-one