network-policies 0.1.0: Restrict traffic to apps' pods
Initial release of helm chart with `NetworkPolicy` resources.
This currently contains only policies to restrict traffic between
webapps' pods but we may add more network policies later (that's why
in the README the installation command install this helm chart
in `default`).

Once this is installed a pod in the `apps-prod` namespace (configurable)
would not be able to talk with other pods in that namespace.

Only traffic allowed will be traffic from the Ingress controller's pod
in the `default` namespace (again, configurable in case these will
run somewhere).

Part of ticket: https://trello.com/c/z0H0xygD/227-prevents-webapp-to-be-able-to-talk-with-each-other