network-policies 0.2.1: allow only traffic from Ingress' pods
once Calico is upgrade in all out clusters we can leverage the support
for both `namespaceSelector` and `podSelector` in `NetworkPolicy`.

This will be more strict and only allow traffic for the Ingress Controller's
pods rather than from all pods in the namepace where the Ingress Controller's
pods run.

ALSO: I've tweaked the `NetworkPolicy`'s indentation to avoid YAML getting
confused between arrays and maps. Arrays are now indented to make it visible
who's the parent and which other properties are at the same level (in this case
`from` and `ports` are at siblings while the pod/namespace selectors are
children of `from`)