unidler v4.0.1: Use strategic merge patch instead of JSONPatch
Use new version of the unidler which uses strategic merge patch instead of
JSONPatch.

Strategic merge is more declarative and more robust. When deleting
a key which is not there it would consider that part of the patch as already
applied instead of raising an error as in the case of JSONpatch.

- [unidler PR](https://github.com/ministryofjustice/analytics-platform-go-unidler/pull/10)
- Part of ticket: https://trello.com/c/vr4LPcde/241-investigate-fix-idling-unidling-problems-caused-by-k8s-cluster-upgrade-new-version-of-k8s