restarter v0.0.1: Initial release
Internal API to restart applications in a given
namespace.

This API doesn't have an Ingress resource so it's
only accessible from within the k8s cluster (at
${RELEASE_NAME}.${NAMESPACE}.svc.cluster.local).

Furthermore, the service account for this API is only able to list and
patch Deployments within the namespace where the chart/API is been
released to.

Part of ticket: https://trello.com/c/zb9HsZca/254-investigate-ways-to-trigger-apps-restart-when-data-changes