idler 2.1.0: Run idler every Tuesday by default
This is to reduce users' waiting time in the morning.

I've updated the default `schedule` as we currently we don't have config
files for this helm chart and it makes sense to keep installation of
this chart simple rather then require a config file.

Ticket: https://trello.com/c/6yyF8WFV